refactor: usePrefersReducedMotion hook, add default value for konami to make it required in foobarData
